# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axgit API",
    "description": "Read-only JSON API over bare Git repositories, plus Smart HTTP (fetch/clone) and an Atom feed.\n\nThere are no mutation endpoints and no authentication: writes happen over SSH on the git-server container, never here.\n\nThis document describes shapes, parameters and status codes. The semantic rules it cannot express — truncation caps, longest-ref matching for the `{ref}`/`{path}` split, merge-commit log simplification, and the exact conditions under which a field is `null` — live in `docs/API.md`, which remains the normative contract.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/repos/{repo}/commits": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "Commit log",
        "description": "Cursor-paginated, newest first. Merge commits survive the `path` filter\nonly when the path differs from **every** parent — an approximation of\n`git log -- <path>` simplification.",
        "operationId": "list_commits",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch, tag, or commit sha; HEAD when absent. Ignored when `cursor` is set.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "path",
            "in": "query",
            "description": "Only commits that changed this file or directory. A path that never\nexisted yields an empty list rather than a 404.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "`next_cursor` from a previous page, walked from **inclusive**. Opaque:\na malformed or unknown value is `400 invalid_param`, not a 404.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Parsed manually so an invalid value yields the JSON `invalid_param`\nenvelope instead of axum's plain-text 400. Never clamped.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            },
            "example": 50
          }
        ],
        "responses": {
          "200": {
            "description": "One page of commits. An empty repository with no `ref` yields an empty page.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; opaque"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitsPage"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "400": {
            "description": "`invalid_param` — bad `limit` or `cursor`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repo}/commits/{sha}": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "Commit detail",
        "description": "Superset of a log entry, plus the full message and the first-parent\ndiffstat. The diffstat has no file cap.",
        "operationId": "get_commit",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "sha",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commit detail. Immutable caching only when `{sha}` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitDetail"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repo}/commits/{sha}/diff": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "Structured commit diff",
        "description": "Unified diff as JSON (file → hunk → line), against the first parent.\nCaps: 1000 rendered lines per file (whole hunks are dropped, never cut) and\n300 files per response; `additions`/`deletions` stay complete either way.",
        "operationId": "get_commit_diff",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "sha",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Restrict the diff to one file (literal match, no globbing). A path that\nthis commit did not touch yields `files: []`, not a 404.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Structured diff. Immutable caching only when `{sha}` is the resolved full sha (then no `ETag`).",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; absent on full-sha requests"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitDiff"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitAuthor": {
        "type": "object",
        "description": "Commit author of `GET /api/v1/repos/{repo}/commits` (docs/API.md).\nThe raw email is never exposed; `email_hash` seeds locally generated avatars.",
        "required": [
          "name",
          "email_hash"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email_hash": {
            "type": "string",
            "description": "sha256 hex of the trimmed, lowercased author email.",
            "example": "b642b4217b34b1e8d3bd915fc65c4452"
          }
        }
      },
      "CommitDetail": {
        "type": "object",
        "description": "Commit detail of `GET /api/v1/repos/{repo}/commits/{sha}` (docs/API.md).\nSuperset of [`CommitInfo`] so the frontend can extend the log entry type.",
        "required": [
          "sha",
          "summary",
          "message",
          "author",
          "committer",
          "authored_at",
          "committed_at",
          "parents",
          "diffstat"
        ],
        "properties": {
          "sha": {
            "type": "string",
            "description": "Full commit sha, even when the request used a branch, tag or short sha."
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Full commit message. `None` for non-utf8 messages."
          },
          "author": {
            "$ref": "#/components/schemas/CommitAuthor"
          },
          "committer": {
            "$ref": "#/components/schemas/CommitAuthor"
          },
          "authored_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "committed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "parents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diffstat": {
            "$ref": "#/components/schemas/DiffStat",
            "description": "First-parent diffstat (docs/API.md)."
          }
        }
      },
      "CommitDiff": {
        "type": "object",
        "description": "Structured diff of `GET /api/v1/repos/{repo}/commits/{sha}/diff` (docs/API.md).",
        "required": [
          "sha",
          "parent",
          "truncated",
          "files"
        ],
        "properties": {
          "sha": {
            "type": "string"
          },
          "parent": {
            "type": [
              "string",
              "null"
            ],
            "description": "First parent the diff was computed against; `null` for a root commit."
          },
          "truncated": {
            "type": "boolean",
            "description": "`true` when files beyond [`MAX_DIFF_FILES`] were omitted."
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDiff"
            }
          }
        }
      },
      "CommitInfo": {
        "type": "object",
        "description": "Log entry of `GET /api/v1/repos/{repo}/commits` (docs/API.md).",
        "required": [
          "sha",
          "summary",
          "author",
          "authored_at",
          "parents"
        ],
        "properties": {
          "sha": {
            "type": "string",
            "description": "Full commit sha."
          },
          "summary": {
            "type": [
              "string",
              "null"
            ],
            "description": "First line of the commit message. `None` for non-utf8 commit messages."
          },
          "author": {
            "$ref": "#/components/schemas/CommitAuthor"
          },
          "authored_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Authordate (RFC 3339). `None` for unrepresentable timestamps.",
            "example": "2026-07-01T14:00:00+09:00"
          },
          "parents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CommitsPage": {
        "type": "object",
        "required": [
          "commits",
          "next_cursor"
        ],
        "properties": {
          "commits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitInfo"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Sha of the first commit of the next page; `null` on the last page.\nPass it back as the `cursor` query parameter."
          }
        }
      },
      "DiffStat": {
        "type": "object",
        "description": "Diffstat of `GET /api/v1/repos/{repo}/commits/{sha}` (docs/API.md).",
        "required": [
          "files",
          "files_changed",
          "total_additions",
          "total_deletions"
        ],
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiffStatFile"
            }
          },
          "files_changed": {
            "type": "integer",
            "minimum": 0
          },
          "total_additions": {
            "type": "integer",
            "minimum": 0
          },
          "total_deletions": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "DiffStatFile": {
        "type": "object",
        "description": "Per-file entry shared by the diffstat and the diff (`#[serde(flatten)]`).",
        "required": [
          "path",
          "old_path",
          "status",
          "additions",
          "deletions",
          "binary"
        ],
        "properties": {
          "path": {
            "type": "string",
            "example": "src/main.rs"
          },
          "old_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Previous path; only set for `renamed`/`copied`."
          },
          "status": {
            "$ref": "#/components/schemas/DiffStatus"
          },
          "additions": {
            "type": "integer",
            "description": "Always the full count, even when the file's hunks were truncated.\nBinary files report 0.",
            "minimum": 0
          },
          "deletions": {
            "type": "integer",
            "description": "Always the full count, even when the file's hunks were truncated.\nBinary files report 0.",
            "minimum": 0
          },
          "binary": {
            "type": "boolean"
          }
        }
      },
      "DiffStatus": {
        "type": "string",
        "description": "How a file changed between the two trees. Rename detection runs with\nlibgit2 defaults (renames only, 50% similarity).",
        "enum": [
          "added",
          "deleted",
          "modified",
          "renamed",
          "copied",
          "typechange"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "One of `repo_not_found`, `ref_not_found`, `path_not_found`,\n`invalid_param`, `read_only`, `internal`.",
            "example": "repo_not_found"
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Internal errors report a generic message; the\ncause only goes to the server log.",
            "example": "repository 'foo' not found"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope shared by every failing endpoint (docs/API.md).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
      "FileDiff": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DiffStatFile"
          },
          {
            "type": "object",
            "required": [
              "truncated",
              "hunks"
            ],
            "properties": {
              "truncated": {
                "type": "boolean",
                "description": "`true` when hunks were dropped past [`MAX_FILE_DIFF_LINES`].\n`additions`/`deletions` still count the full change."
              },
              "hunks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Hunk"
                },
                "description": "Empty for binary files."
              }
            }
          }
        ]
      },
      "Hunk": {
        "type": "object",
        "required": [
          "header",
          "old_start",
          "old_lines",
          "new_start",
          "new_lines",
          "lines"
        ],
        "properties": {
          "header": {
            "type": "string",
            "description": "`@@ -a,b +c,d @@ <context>` line without the trailing newline.",
            "example": "@@ -1,2 +1,2 @@"
          },
          "old_start": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "old_lines": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "new_start": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "new_lines": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            }
          }
        }
      },
      "Line": {
        "type": "object",
        "required": [
          "origin",
          "content",
          "old_lineno",
          "new_lineno"
        ],
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/LineOrigin"
          },
          "content": {
            "type": "string",
            "description": "Line content without the trailing newline (lossy utf-8)."
          },
          "old_lineno": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "`null` on added lines.",
            "minimum": 0
          },
          "new_lineno": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "`null` on deleted lines.",
            "minimum": 0
          }
        }
      },
      "LineOrigin": {
        "type": "string",
        "description": "Which side of the diff a line belongs to. Other libgit2 origins (EOF\nnewline markers and the like) are dropped rather than reported.",
        "enum": [
          " ",
          "+",
          "-"
        ]
      },
...
    }
  },
...
}